Backed out changeset 14b092e7a974 (bug 1635762) for causing failures in browser_Telemetry_timestamp_test.js
CLOSED TREE

+++ tools/fuzzing/libfuzzer/clone_libfuzzer.sh
@@ -1,35 +1,31 @@
-#!/bin/bash -e
+#!/bin/bash
 
 # Optionally get revision from cmd line
-[ $1 ] && REVISION=$1 || REVISION=c815210013f27cfac07d6b53b47e8ac53e86afa3
+# Changelog: https://reviews.llvm.org/source/compiler-rt/history/compiler-rt/trunk/lib/fuzzer/
+[ $1 ] && REVISION=$1 || REVISION=356365
 
 mkdir tmp
-git clone --single-branch --no-checkout --shallow-since "2019-03-01" https://github.com/llvm/llvm-project tmp
+svn co -qr $REVISION https://llvm.org/svn/llvm-project/compiler-rt/trunk tmp || exit
 
-(cd tmp && git reset --hard $REVISION)
+if [ $1 ]; then
+  # libFuzzer source files
+  CPPS=($(ls -rv tmp/lib/fuzzer/*.cpp))
+  CPPS=(${CPPS[@]##*/})
+  CPPS=(${CPPS[@]##FuzzerMain*}) # ignored
 
-# libFuzzer source files
-CPPS=($(ls -rv tmp/compiler-rt/lib/fuzzer/*.cpp))
-CPPS=(${CPPS[@]##*/})
-CPPS=(${CPPS[@]##FuzzerMain*}) # ignored
+  # Update SOURCES entries
+  sed -e "/^SOURCES/,/^]/ {/'/d}" -i moz.build
+  for CPP in ${CPPS[@]}; do sed -e "/^SOURCES/ a \\\t'${CPP}'," -i moz.build; done
 
-# Update SOURCES entries
-sed -e "/^SOURCES/,/^]/ {/'/d}" -i moz.build
-for CPP in ${CPPS[@]}; do sed -e "/^SOURCES/ a \\    '${CPP}'," -i moz.build; done
-
-# Remove previous files
-rm *.{cpp,h,def}
+  # Remove previous files
+  rm *.{cpp,h,def}
+fi
 
 # Copy files
-cp tmp/compiler-rt/lib/fuzzer/*.{cpp,h,def} .
-
-# Apply local patches
-for patch in patches/*.patch
-do
-  patch -p4 < $patch
-done
+cp tmp/lib/fuzzer/*.{cpp,h,def} .
 
 # Remove the temporary directory
 rm -Rf tmp/
 
-echo "Updated libFuzzer to ${REVISION}"
+[ $1 ] && echo "Updated libFuzzer to ${REVISION}"
+
